Eu te amo Thomas Borregaard Sørensen!!
Eu te amo Mummi!! Feliz 62 anos!!
ca2 draw2d implementation and draw2d cairo implementation contribution!! direct2d in Windows 8 Desktop!!

Eu te amo Thomas Borregaard Sørensen!!

Mummi contribution!!!! Feliz 62 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU 4 core 8 threads @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- 10Mpbs Download and 2Mpbs Upload NET Virtua Internet with LG 23´´ monitor, Microsoft Comfort Curve 2000 v1.0 Wired Keyboard and Microsoft Comfort Optical Wired Mouse 3000 at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

｛
｝
                        -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/draw2d_gdi/draw2d_gdi_graphics.cpp
@@ -1582,7 +1582,7 @@ namespace draw2d_gdi
 
          ::SetTextColor(m_hdc, RGB(GetRValue(brush.m_cr), GetGValue(brush.m_cr), GetBValue(brush.m_cr)));
 
-         return ::TextOutW(get_handle1(), x, y, wstr, wstr.length()) != FALSE;
+         return ::TextOutW(get_handle1(), x, y, wstr, (int) wstr.length()) != FALSE;
 
       }
       else
@@ -1590,7 +1590,7 @@ namespace draw2d_gdi
 
          SIZE size;
 
-         if(!::GetTextExtentPoint32W(get_handle2(), wstr, (int)wstr.get_length(), &size))
+         if(!::GetTextExtentPoint32W(get_handle2(), wstr, (int) wstr.get_length(), &size))
          {
 
             return false;
@@ -1626,7 +1626,7 @@ namespace draw2d_gdi
 
          ::SetBkMode((HDC) pdib->get_graphics()->get_os_data(), TRANSPARENT);
 
-         ::TextOutW((HDC) pdib->get_graphics()->get_os_data(), 0, 0, wstr, wstr.length());
+         ::TextOutW((HDC) pdib->get_graphics()->get_os_data(), 0, 0, wstr, (int) wstr.length());
 
          GDI_DIB(pdib)->process_blend(&brush, x, y, m_ealphamode, m_pdib);
 
@@ -1655,7 +1655,7 @@ namespace draw2d_gdi
    bool graphics::TextOut(int x, int y, const string & str)
    { 
 
-      return TextOut(x, y, str, str.get_length());
+      return TextOut(x, y, str, (int) str.get_length());
 
    }
 
@@ -1723,7 +1723,7 @@ namespace draw2d_gdi
    int graphics::draw_text(const string & str, LPRECT lpRect, UINT nFormat)
    { 
 
-      return draw_text(str, str.get_length(), lpRect, nFormat);
+      return draw_text(str, (int) str.get_length(), lpRect, nFormat);
 
    }
 
@@ -1733,7 +1733,7 @@ namespace draw2d_gdi
 
       wstring wstr(string(lpszString, nCount));
 
-      return DrawTextExW(get_handle(), (wchar_t *) (const wchar_t *)  wstr, wstr.get_length(), lpRect, nFormat, lpDTParams);
+      return DrawTextExW(get_handle(), (wchar_t *) (const wchar_t *)  wstr, (int) wstr.get_length(), lpRect, nFormat, lpDTParams);
 
    }
 
@@ -1741,7 +1741,7 @@ namespace draw2d_gdi
    int graphics::draw_text_ex(const string & str, LPRECT lpRect, UINT nFormat, LPDRAWTEXTPARAMS lpDTParams)
    { 
 
-      return draw_text_ex((char *) (const char *) str, str.get_length(), lpRect, nFormat, lpDTParams);
+      return draw_text_ex((char *) (const char *) str, (int) str.get_length(), lpRect, nFormat, lpDTParams);
 
    }
 
@@ -1896,7 +1896,7 @@ namespace draw2d_gdi
 
       ASSERT(get_handle1() != NULL);
 
-      return ::GetCharacterPlacement(get_handle1(), (const char *)str, str.get_length(), nMaxExtent, lpResults, dwFlags);
+      return ::GetCharacterPlacement(get_handle1(), (const char *)str, (int) str.get_length(), nMaxExtent, lpResults, dwFlags);
 
    }
